Move v2 routes to v2 module
