+++ src/main.rs
@@ -366,8 +366,8 @@ async fn main() -> std::io::Result<()> {
             .app_data(web::Data::new(payouts_queue.clone()))
             .app_data(web::Data::new(ip_salt.clone()))
             .wrap(sentry_actix::Sentry::new())
-            .configure(routes::v1_config)
-            .configure(routes::v2_config)
+            .configure(routes::v1::config)
+            .configure(routes::v2::config)
             .service(routes::index_get)
             .service(routes::health_get)
             .service(web::scope("maven").configure(routes::maven_config))
