+++ src/routes/v2/health.rs
