+++ src/routes/v2/statistics.rs
